Eliminated cross command imports (except one)
