+++ cdv/cmds/clsp.py
@@ -15,33 +15,11 @@ from clvm_tools.binutils import disassemble, assemble
 from chia.types.blockchain_format.program import Program
 from chia.util.byte_types import hexstr_to_bytes
 
+from cdv.cmds.util import parse_program, append_include
+
 @click.group("clsp", short_help="Commands to use when developing with chialisp")
 def clsp_cmd():
     pass
-
-def append_include(search_paths):
-    if search_paths:
-        list(search_paths).append("./include")
-    else:
-        return ['./include']
-
-def parse_program(program: str, include=[]):
-    if '(' in program:
-        prog = Program.to(assemble(program))
-    elif '.' not in program:
-        prog = Program.from_bytes(hexstr_to_bytes(program))
-    else:
-        with open(program, "r") as file:
-            filestring = file.read()
-            if '(' in filestring:
-                # TODO: This should probably be more robust
-                if re.compile('\(mod\s').search(filestring):
-                    prog = Program.to(compile_clvm_text(filestring, append_include(include)))
-                else:
-                    prog = Program.to(assemble(filestring))
-            else:
-                prog = Program.from_bytes(hexstr_to_bytes(filestring))
-    return prog
 
 @clsp_cmd.command("build", short_help="Build all specified CLVM files (i.e mypuz.clsp or ./puzzles/*.clsp)")
 @click.argument("files", nargs=-1, required=True, default=None)
